Conflicts:
	Source/WTF/wtf/MathExtras.h
	Source/cmake/OptionsCommon.cmake
	Tools/Scripts/webkitdirs.pm

+++ Source/JavaScriptCore/heap/MarkedSpace.cpp
@@ -90,6 +90,7 @@ MarkedSpace::MarkedSpace(Heap* heap)
         destructorAllocatorFor(cellSize).init(heap, this, cellSize, true, false);
     }
 
+    m_largeAllocator.init(heap, this, 0, true, false);
     m_structureAllocator.init(heap, this, WTF::roundUpToMultipleOf(32, sizeof(Structure)), true, true);
 }
 
@@ -127,6 +128,7 @@ void MarkedSpace::resetAllocators()
         destructorAllocatorFor(cellSize).reset();
     }
 
+    m_largeAllocator.reset();
     m_structureAllocator.reset();
 }
 
@@ -144,16 +146,17 @@ void MarkedSpace::reapWeakSets()
 void MarkedSpace::canonicalizeCellLivenessData()
 {
     for (size_t cellSize = preciseStep; cellSize <= preciseCutoff; cellSize += preciseStep) {
-        allocatorFor(cellSize).zapFreeList();
-        destructorAllocatorFor(cellSize).zapFreeList();
+        allocatorFor(cellSize).canonicalizeCellLivenessData();
+        destructorAllocatorFor(cellSize).canonicalizeCellLivenessData();
     }
 
     for (size_t cellSize = impreciseStep; cellSize <= impreciseCutoff; cellSize += impreciseStep) {
-        allocatorFor(cellSize).zapFreeList();
-        destructorAllocatorFor(cellSize).zapFreeList();
+        allocatorFor(cellSize).canonicalizeCellLivenessData();
+        destructorAllocatorFor(cellSize).canonicalizeCellLivenessData();
     }
 
-    m_structureAllocator.zapFreeList();
+    m_largeAllocator.canonicalizeCellLivenessData();
+    m_structureAllocator.canonicalizeCellLivenessData();
 }
 
 bool MarkedSpace::isPagedOut(double deadline)
@@ -168,6 +171,9 @@ bool MarkedSpace::isPagedOut(double deadline)
             return true;
     }
 
+    if (m_largeAllocator.isPagedOut(deadline))
+        return true;
+
     if (m_structureAllocator.isPagedOut(deadline))
         return true;
 
@@ -178,7 +184,12 @@ void MarkedSpace::freeBlock(MarkedBlock* block)
 {
     allocatorFor(block).removeBlock(block);
     m_blocks.remove(block);
-    m_heap->blockAllocator().deallocate(MarkedBlock::destroy(block));
+    if (block->capacity() == MarkedBlock::blockSize) {
+        m_heap->blockAllocator().deallocate(MarkedBlock::destroy(block));
+        return;
+    }
+
+    MarkedBlock::destroy(block).deallocate();
 }
 
 void MarkedSpace::freeOrShrinkBlock(MarkedBlock* block)
